# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Reservas de Restaurantes",
        "description": "API para la gestión de clientes, mesas y reservas.",
        "contact": {
            "name": "Vladimir Faundez",
            "email": "v.faundezh@gmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International",
            "url": "https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "description": "Retrieve a list of all customers.",
                "operationId": "getAllCustomers",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/SearchParam"
                    },
                    {
                        "$ref": "#/components/parameters/CustomerSortByParam"
                    },
                    {
                        "$ref": "#/components/parameters/SortDirParam"
                    },
                    {
                        "$ref": "#/components/parameters/DateFromParam"
                    },
                    {
                        "$ref": "#/components/parameters/DateToParam"
                    },
                    {
                        "$ref": "#/components/parameters/DateFieldParam"
                    },
                    {
                        "$ref": "#/components/parameters/CustomerIncludeParam"
                    },
                    {
                        "$ref": "#/components/parameters/PageParam"
                    },
                    {
                        "$ref": "#/components/parameters/PerPageParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/CustomerSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "description": "Store a new customer in the system.",
                "operationId": "storeCustomer",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreCustomerRequest"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/CustomerSuccessId"
                    },
                    "400": {
                        "$ref": "#/components/responses/CustomerInvalidRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/{customer}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer by ID",
                "description": "Retrieve a specific customer by their ID.",
                "operationId": "getCustomerById",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID of the customer to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/CustomerIncludeParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/CustomerSuccessId"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "description": "Update an existing customer in the system.",
                "operationId": "updateCustomer",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID of the customer to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateCustomerRequest"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/CustomerSuccessId"
                    },
                    "400": {
                        "$ref": "#/components/responses/CustomerInvalidRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer",
                "description": "Remove a specific customer from the system.",
                "operationId": "deleteCustomer",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID of the customer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer deleted successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Customer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@mail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+56912345678"
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ServerErrorResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Internal server error"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "CustomerSuccess": {
                "description": "Successful customer operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "CustomerSuccessId": {
                "description": "Successful customer operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                },
                                "data": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "CustomerInvalidRequest": {
                "description": "Invalid request data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Invalid request"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "email"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "The email field is required."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFoundError": {
                "description": "Reservation not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
            "UnauthenticatedError": {
                "description": "Unauthenticated user",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthenticatedResponse"
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Unauthorized action",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthorizedResponse"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ServerErrorResponse"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "CustomerSortByParam": {
                "name": "sort_by",
                "in": "query",
                "description": "Field to sort customers by.",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "id",
                    "enum": [
                        "id",
                        "name",
                        "email",
                        "phone"
                    ]
                }
            },
            "CustomerIncludeParam": {
                "name": "include",
                "in": "query",
                "description": "Comma-separated list of relationships to include.",
                "required": false,
                "schema": {
                    "type": "string",
                    "example": "reservations"
                }
            },
            "SearchParam": {
                "name": "search",
                "in": "query",
                "description": "Search term for textual filtering.",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
...
            "SortDirParam": {
                "name": "sort_dir",
                "in": "query",
                "description": "Sort direction.",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "asc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            },
            "DateFromParam": {
                "name": "date_from",
                "in": "query",
                "description": "Filter records from this date (inclusive). Format: YYYY-MM-DD",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "date"
                }
            },
            "DateToParam": {
                "name": "date_to",
                "in": "query",
                "description": "Filter records up to this date (inclusive). Format: YYYY-MM-DD",
                "required": false,
                "schema": {
                    "type": "string",
                    "format": "date"
                }
            },
            "DateFieldParam": {
                "name": "date_field",
                "in": "query",
                "description": "Field to apply the date filter on.",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "created_at"
                }
            },
...
            "PageParam": {
                "name": "page",
                "in": "query",
                "description": "Current page number.",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "PerPageParam": {
                "name": "per_page",
                "in": "query",
                "description": "Number of records per page.",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 15
                }
            }
        },
        "requestBodies": {
            "StoreCustomerRequest": {
                "description": "Data required to create a customer",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "email",
                                "phone"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "John Doe"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "johndoe@example.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "+123456789"
                                },
                                "registrationDate": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2024-01-01T12:00:00Z"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateCustomerRequest": {
                "description": "Data to update a customer",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "John Doe"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "johndoe@example.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "+123456789"
                                },
                                "registrationDate": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2024-01-01T12:00:00Z"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT Authorization header usando el esquema Bearer.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}